Separated viewModels for tasks and timer.

+++ app/src/main/java/com/fylora/pomodorotimer/data/entity/TaskEntity.kt
@@ -9,8 +9,8 @@ data class TaskEntity(
     val title: String,
     val workTime: Long,
     val dueDate: Long,
-    val totalSessions: Int,
-    val currentSessions: Int,
+    val totalSessions: Double,
+    val currentSessions: Double,
 
     @PrimaryKey
     val id: Int? = null
